Fix issue #33 do not remove LDAP users from CF
Use the same consistent key when accessing the space users map.

When using an LDAP configuration with an origin of something besides 'ldap', the space users map key was getting set to the user's lower case LDAP email address. This allowed the updateLdapUser func to find the user in the existing map of space users, however when it went to remove that key from the map of space users it used the LDAP user's user id instead of email.