Reduce config duplication for orgs
The config package is supposed to provide the utilities for expressing the desired configuration of a CF installation.  It exposes interfaces for reading and writing to "the cf-mgmt config."

The fact that we use a directory of YAML files as our config repository should be an implementation detail of the config package.  For example, we should be able to implement a different config manager that uses a database or a different file format, and the rest of the application shouldn't care.

The fact that we use YAML leaks outside of the config package though.  Configuration structs are duplicated in the config package and the other {organization|space} packages.  The structs with YAML tags are not in the config package, but in the other packages.  The abstraction is broken.

Start to rectify this by reducing duplication and moving config structures into the config package.  We're inverting the dependency, with the end goal being that the rest of the application depends on the config package, rather than the current state where the config package depends on the rest of the application.

This commit moves org config into the org package.  We're using Go 1.9 type aliases here to enable incremental refactoring, so this won't build on Go < 1.9 until we're complete.