Move code that reads YAML config into the config package
Thus far the config.Manager interface exposed an API for updating the config, but no way to read the config.

The organization and space packages themselves had to know about the implementation details of the YAML config, and were forced to read the files themselves.

This commit breaks the config interfaces into an Updater (formerly Manager) and a Reader.  The types in the organization and space packages can now use a config.Reader instead of reaching into the YAML on their own.

Start with moving the ability to read all orgs from the org manager to the config reader.