start modifying isolation segments
Add new package isosegments that will be used to create and enable access to isolation segments.

The public API for this package will include:
- Ensure: verify that all isolation segments exist [DONE]
- Entitle: verify that orgs are entitled to the required isolation segments [DONE]
- UpdateOrgs: verify that orgs default iso segment is set [TODO]
- UpdateSpaces: verify that spaces default iso segment is set [TODO]

Also note - as part of this commit we introduce code.cloudfoundry.org/cli/api/cloudcontroller.  This is the go package used by the CLI.  It is my hope that after the isolation segments experiment is complete that we can update the rest of the code base to use this package and do away with our custom cloudcontroller package.  This would be less code to maintain in cf-mgmt and would ensure consistency between this tool and the CLI.