Update integration tests to use the cf CLI
The integration tests now use the cf-mgmt CLI to invoke some behavior, and verify the outcome using the cf CLI.  The tests are a little slower, but we're not using our own code to verify our code, and there's a lot less setup to do.

Add isolation-segments tests to integration tests.

Mention in the README that integration tests require cf CLI.