182663012 test against tas before merging (#297)
* Run test-against-tas jobs react to PRs not merges

[#182663012]

Previously, these jobs triggered after the PR was merged.
This approach doesn't present any benefits and you might
discover that the PR included breaking changes immediately
after mergin.

Why not test the PRs before merging then? This commit is
the first step towards that goal.

* Extract `merge` tasks to its own job

[#182663012]

To be able to describe that the merge should only occur after
all tests have run we need to extract this logic to its own job

* Extract claim and unclaim tasks to their own jobs

[#182663012]

This allows more granular control while debugging. It also
provides better visibility about what the pipeline is doing.

* Improve pipeline legibility

[#182663012]

This change doesn't change pipeline semantics that much but
the graph that Concourse draws is clearer

* Get rid of useless step and resource

[#182663012]

According to some docs from an old commit the reason behind this step was:
> This is just so that we can trigger the job everytime there's a new TAS
> version.  It will download the tile... which is not quite nice in terms of
> execution time and resourse consumption. Maybe there's a better way.
https://github.com/vmware-tanzu-labs/cf-mgmt/blob/ea7d21a799196b7a31d1231abd061dbf7c7cec8a/ci/pipeline.yml#L35-L38

Since we want to trigger on PR changes not on changes to pivnet
there is no reason to keep this resource and logic around anymore

* Extract claim and unclaim tasks also for cf-deployment

[#182663012]

This allows more granular control while debugging. It also
provides better visibility about what the pipeline is doing.

* Add or rename missing resources in merge job

[#182663012]